[SPARK-31831] put mocks in hive version subdirectory
